Snapshots for events and snapshots were missing one entry as iterator toToken is exclusive Changed procedure for deleting files to ensure that buffers are closed before removing Do not try to include event and snapshot data in snapshots for _admin context Delete applications/users only in _admin context Added lastAppliedSnapshotSequence to failure message Changed default values for cleanupDelay properties as the unit is seconds (not millisecs) DefaultCurrentConfiguration had iterator that could start before first index in files Creation of GrpcRaftController failed (probably only when management.endpoint.health.show-details=always)

+++ axonserver-enterprise/src/main/java/io/axoniq/axonserver/enterprise/cluster/snapshot/ApplicationSnapshotDataStore.java
@@ -74,7 +74,9 @@ public class ApplicationSnapshotDataStore implements SnapshotDataStore {
 
     @Override
     public void clear() {
-        applicationController.deleteByContext(context);
+        if( adminContext) {
+            applicationController.deleteByContext(context);
+        }
     }
 
     private SerializedObject toSerializedObject(io.axoniq.axonserver.grpc.internal.Application application) {
